pyproject.toml
[tool.coverage.run]
omit = [
    "ambuda/scripts/*",
    "ambuda/seed/**",
]

[tool.poetry]
name = "ambuda"
version = "0.1.0"
description = "An open Sanskrit library"
authors = ["Arun Prasad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.8.0"
amqp = "5.1.1"
Babel = "^2.12"
"backports.functools-lru-cache" = "1.6.4"
bcrypt = "4.0.0"
click = "8.1.3"
conllu = "4.5.2"
email-validator = "1.2.1"
fabric = "2.7.0"
Flask = "2.1.2"
Flask-Admin = "1.6.0"
Flask-Babel = "3.0.1"
Flask-Bcrypt = "1.0.1"
Flask-Login = "0.6.1"
Flask-Mail = "0.9.1"
Flask-WTF = "1.0.1"
google-api-core = "2.8.2"
google-auth = "2.9.0"
google-cloud-vision = "3.1.1"
googleapis-common-protos = "1.56.3"
greenlet = "1.1.2"
gunicorn = "20.1.0"
indic-transliteration = "2.3.31"
Jinja2 = "3.1.2"
kombu = "5.2.4"
linkify-it-py = "2.0.0"
lxml = "4.9.1"
Mako = "1.2.2"
markdown-it-py = "2.1.0"
MarkupSafe = "2.1.1"
mypy-extensions = "0.4.3"
Pillow = "9.1.1"
PyMuPDF = "1.20.2"
python-dateutil = "2.8.2"
python-dotenv = "0.20.0"
python-slugify = "6.1.2"
pytz = "2022.7.0"
regex = "2022.6.2"
requests = "2.27.1"
sentry-sdk = "1.6.0"
SQLAlchemy = "1.4.37"
Werkzeug = "2.1.2"
WTForms = "3.0.1"
wtforms-sqlalchemy = "^0.3"


[tool.poetry.dev-dependencies]
# Unit tests
pytest = "^7.1.2"
pytest-cov = "4.0.0"
# Linting / formatting
black = "^22.3.0"
ruff = "0.0.260"
# Documentation
Sphinx = "5.0.2"
sphinx-rtd-theme = "1.0.0"
sphinxcontrib-applehelp = "1.0.2"
sphinxcontrib-devhelp = "1.0.2"
sphinxcontrib-htmlhelp = "2.0.0"
sphinxcontrib-jsmath = "1.0.1"
sphinxcontrib-qthelp = "1.0.3"
sphinxcontrib-serializinghtml = "1.1.5"


[build-system]
requires = ["poetry-core", "freetype"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# https://beta.ruff.rs/docs/rules/
select = [
    "B",     # flake8-bugbear
    "F",     # pyflakes
    "N818",  # error-suffix-on-exception-name
    "UP",    # pyupgrade
    "W",     # pycodestyle
    "I001",  # isort
]
ignore = [
    "B007",  # unused-loop-control-variable
    "B905"   # zip-without-explicit-strict
]
